requirements.txt
aiohttp>=3.9.0
websockets>=12.0
pandas>=2.1.0
numpy>=1.26.0
pandas-ta>=0.3.14b
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiosqlite>=0.19.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
